+++ script-test/src/test/java/com/aplana/sbrf/taxaccounting/form_template/ndfl/primary_rnu_ndfl/v2016/PrimaryRnuNdflScriptCheckTest.java
@@ -1318,7 +1318,7 @@ public class PrimaryRnuNdflScriptCheckTest extends DeclarationScriptTestBase {
         NdflPersonIncome ndflPersonIncome = new NdflPersonIncome();
 
         ndflPersonIncome.setNdflPersonId(ndflPersonId);
-        ndflPersonIncome.setRowNum(BigDecimal.valueOf(rowNum));
+        ndflPersonIncome.setRowNum(rowNum);
 
         ndflPersonIncome.setIncomeCode(column4);
         ndflPersonIncome.setIncomeType(column5);
